+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New UI script runner feature. Launches _bellplay~_ if needed and loads script via OSC.
+- GUI: currently loaded script tab now displays emoji.
 
 ### Fixed